requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
passlib
python-dotenv
langchain-groq
